Rename standalone to all-in-one
