+++ cmd/all-in-one/sampling_strategies.json
@@ -1,0 +1,6 @@
+{
+  "default_strategy": {
+    "type": "probabilistic",
+    "param": 1
+  }
+}
