--- cmd/standalone/sampling_strategies.json
@@ -1,6 +1,0 @@
-{
-  "default_strategy": {
-    "type": "probabilistic",
-    "param": 1
-  }
-}
