+++ scripts/travis/build-all-in-one-image.sh
@@ -10,7 +10,7 @@ make build-all-in-one-linux
 
 export REPO=jaegertracing/all-in-one
 
-docker build -f cmd/standalone/Dockerfile -t $REPO:latest .
+docker build -f cmd/all-ine-one/Dockerfile -t $REPO:latest .
 export CID=$(docker run -d -p 16686:16686 -p 5778:5778 $REPO:latest)
 make integration-test
 docker kill $CID
